Rework callbacks to module
